adaptivedemux: minimal HTTP context support
The uridownloader is now querying the source element for an HTTP
context, which stores session data (cookies only for now), and reusing
the data when fetching data over HTTP. Additionally the context is set
on adaptivedemux, which allows it to also properly use session data
when downloading fragments.

https://bugzilla.gnome.org/show_bug.cgi?id=726314

Reworked by Enrique Ocaña González <eocanha@igalia.com on May 23 2018
in the upgrade from GStreamer 1.10.4 to 1.14.1.