Fix HLS live stream issues with http://cdn.metrological.com/hls/master.m3u8
This live stream has index{0-3}.ts blocks which repeat forever. The
first issue is that the fetch of an updated playlist takes too much and
the playback stalls. This has been fixed by increasing
GST_M3U8_LIVE_MIN_FRAGMENT_DISTANCE from 3 to 4, which gives more
buffer at the expense of latency in the stream (it's less realtime).

The second issue was that index0.ts starts with a non-zero Program
Clock Reference (PCR, the timing used in MPEG TS). When the blocks wrap
around from index3.ts to index0.ts, the PCR reset code detects a
non-zero gap and screws the timestamps. This is fixed by correcting the
out_time, but requires to disable PCR wraparound management, which
seems a bit fishy.

WARNING: This patch is experimental and can produce regressions!